Change path list elements to a tuple
Before, path list elements were just the nodes
which are part of the same path. However, when
exploring the found path, there was no way to
know how many iterations a loop did since there
was not any information about its WCEC or number
of iterations at that path.

Now, the nodes and its WCEC are store in path
list. So, it is easier to know the number of
iterations of each loop.