Working with ready queue in simulation manager
Ready queue is used now to know which task will
be executed right after the end of the current
task. Each task that ends, it is added to ready
queue again schedule according to its period.

Stop constraint was also defined as the deadline
of the less priority task. In this way, the less
priority task will run only once.