Fix time bugs and improve results information
Simulation time was fixed since jitter was not
reflect in some test cases.

Result informations are more detailed now
including start and end times. A header was also
added to file result.

Result file is now being checked if it exists. If
so, it is removed and created a new one.