Overhead and Energy Consumption Comparison
Overhead - it should be add to cost every time it
happens and not only when frequency changes.

Energy Consumption Comparison - mine approach was
being compared with the highest frequency with
the same number of cycles. However, since mine
approach has overheads, it was not fair to check
energy consumption with the same number of
cycles. So, get path cost without overheads to
the highest frequency.